Fix bug in DMA read code.
I had added a feature in the driver that I never got around to in the
Chisel generator. Naturally, this caused weirdness to happen and the DMA
would unconditionally time-out. I go back to the old way of checking for
DMA completion, which is admittedly imperfect but seems to work OK in
practice.